requirements.txt
hachoir==3.1.1
mongoengine==0.23.1
# face-recognition==1.3.0
numpy==1.20.2
mongoengine==0.23.1
pillow==8.2.0
scipy==1.6.3
tqdm==4.59.0
click==8.0.1
geocoder==1.38.1
what3words==3.1.1
folium==0.12.1
airtable-python-wrapper==0.15.2
phonenumbers==8.12.24
blinker==1.4
flask-mongoengine==1.0.0
flask-bootstrap==3.3.7.1
flask_nav==0.6
deepface==0.0.53
python-slugify==5.0.2
sentry_sdk
flask-login==0.5.0
wtforms==2.3.3
bcrypt==3.2.0
flask-bcrypt==0.7.1
